microraiden/stale_state_attack/geth-cluster/requirements.txt
eth-abi==0.5.0
rlp>=0.4.3,<0.6.0
ethereum<2.0.0,>=1.6.1
web3==v4.0.0-beta.4
py-geth==v2.0.1
click
